#8318 RG service negative
git-svn-id: http://svn.kolmisoft.com/mor/gui/branches/12@39808 19941b1f-7a74-413c-ab52-0de3e92e9a87

+++ app/models/service.rb
@@ -35,8 +35,8 @@ class Service < ActiveRecord::Base
   end
   
   def handle_negatives 
-    write_attribute(:price, 0) if read_attribute(:price).to_i < 0 
-    write_attribute(:selfcost_price, 0) if read_attribute(:selfcost_price).to_i < 0 
+    write_attribute(:price, 0) if read_attribute(:price).to_f < 0 
+    write_attribute(:selfcost_price, 0) if read_attribute(:selfcost_price).to_f < 0 
   end 
 
   # converted attributes for user in current user currency
